Add files via upload
The fluctuation in stock prices for companies – Facebook, Amazon, Apple, Netflix, Google over the first 200 days. Each line represents the high stock prices of a specific company, while the shaded areas between the lines depict the range between the high and low stock prices.
The candlestick graph depicts the technical analysis of equity and currency price patterns. It determines possible price movement based on patterns along with the opening price, closing price, high and low of that time period. The rectangular body is colored with a dark color red for a drop in price and a light color green for a price increase.